Connection with building
building does not receive the power from plb

+++ test code/PLB/ESP industry/lib/HAL/IBuilding.h
@@ -5,7 +5,7 @@ class IBuilding
 {
 public:
     virtual ~IBuilding() = default;
-    virtual int calculateSolarPower() = 0;
+    virtual float calculateSolarPower() = 0;
     virtual void charge(float power) = 0;
 };
